+++ concrete_datastore/admin/admin_site.py
@@ -102,7 +102,10 @@ def get_admin_site():
         def get_app_list(self, request):
             app_dict = self._build_app_dict(request)
             if not app_dict:
-                return []
+                #: This case should never occur, due to the admin site
+                #: permssions set to "superuser".
+                return []  # pragma: no cover
+
             # Sort the apps alphabetically.
             app_list = sorted(
                 app_dict.values(), key=lambda x: x['name'].lower()
@@ -153,7 +156,9 @@ def get_admin_site():
 
         def index(self, request, extra_context=None, *args, **kwargs):
             if settings.USE_CORE_AUTOMATION:
-                extra_context = {
+                #: Ignore the following for coverage because it needs
+                #: real import and real packages
+                extra_context = {  # pragma: no cover
                     'use_core_automation': True,
                     'target_admin_view': reverse('coreAdmin:index'),
                     'target_admin_view_name': 'Core Admin',