+++ tests/tests_api_v1_1/test_api_v1_1_token_expiry.py
@@ -280,3 +280,34 @@ class AuthTestCase(APITestCase):
             resp.status_code, status.HTTP_401_UNAUTHORIZED, msg=resp.content
         )
         self.assertEqual(Project.objects.count(), 2)
+
+    def test_token_with_blocked_user(self):
+        url = '/api/v1.1/auth/login/'
+        url_projects = '/api/v1.1/project/'
+
+        # Login to generate token
+        resp = self.client.post(
+            url, {"email": 'johndoe@netsach.org', "password": "plop"}
+        )
+        self.assertEqual(
+            resp.status_code, status.HTTP_200_OK, msg=resp.content
+        )
+        token_key = resp.data['token']
+        self.user.is_active = False
+        self.user.save()
+
+        # CREATE a project with the user blocked
+        resp = self.client.post(
+            url_projects,
+            {
+                "name": "Projects1",
+                "description": "description de mon projet",
+                "skills": [],
+                "members": [],
+            },
+            HTTP_AUTHORIZATION='Token {}'.format(token_key),
+        )
+        self.assertEqual(
+            resp.status_code, status.HTTP_401_UNAUTHORIZED, msg=resp.content
+        )
+        self.assertEqual(Project.objects.count(), 0)