+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 ### Changed
 
 - The distance filters now support the lookups `__distance_lte`, `__distance_lt`, `__distance_gte`, `__distance_gt`, `__distance_range`, `__distance_range!`. The bare `__distance` is no more supported
+- Maintenance mode flag in the status view if there is a core app
+
+### Changed
+
+- Improve unittests
 
 ### Removed